Move several components into their own directory
